fix(backend): close AUDIT-B8 — the config/dispatcher half
WORKER_URLS defaulted to http://worker:9091; the worker listens on 8000
everywhere else. Nothing caught it because compose always sets WORKER_URLS
explicitly and every test in WorkerDispatcherServiceTest sets the field by
reflection, so the @Value default was the one value never exercised.

WorkerDispatcherService's @PostConstruct init() re-read WORKER_API_SECRET_FILE
off disk, duplicating DockerSecretsEnvironmentPostProcessor, which already turns
every *_FILE env var into the matching property with addFirst. Removed. This
bullet has no red-green and that is not an oversight: it is duplication removal,
not a defect, so no test can go red when it is restored. Its two tests covered
the deleted method and are replaced by one pinning the contract the deletion
depends on — that the bean resolves its secret from the WORKER_API_SECRET
property. Net -1 test.

jwtExpirationMs is now a long. The entry predicts an expiry wrapping into the
past; what actually happens with a 30-day value is that the token comes back
unparseable, because exp lands far enough back that parsing throws before any
claim is returned. Every token minted would be born expired.

JwtAuthFilter:58 was filed wrong and this was verified, not argued. The claim is
that the {} placeholder swallows the throwable so no stack trace is logged. The
inherited logger is a commons-logging Log from GenericFilterBean, whose only
two-arg overload is error(Object, Throwable) — so the throwable was always
attached. Re-running the test with the original line and the {} assertion
removed shows getThrowableProxy() non-null. The real defect is the inverse:
commons-logging does not interpolate, so {} printed literally. The class now has
its own SLF4J logger, which removes the ambiguity that caused the misreading.

Also fixes a PMD violation the previous commit introduced: deleting
regionCallback's discarded resolveNotificationContext(imageId) left imageId
unused. mvn -o clean test does not run PMD; CI runs mvn clean verify, which
does. The gate should be verify.

Backend 412, mvn -o clean verify, full suite. Closes AUDIT-B8.

+++ backend/src/main/java/com/manga/library/config/JwtUtils.java
@@ -16,8 +16,13 @@ public class JwtUtils {
   @Value("${jwt.secret}")
   private String jwtSecret;
 
+  /**
+   * AUDIT-B8: a {@code long}. As an {@code int} this capped at ~24.8 days, and the overflow did not
+   * fail loudly — {@code new Date(now + jwtExpirationMs)} with a wrapped negative value produces an
+   * expiry in the *past*, so every token minted would be born already expired.
+   */
   @Value("${jwt.expirationMs}")
-  private int jwtExpirationMs;
+  private long jwtExpirationMs;
 
   private SecretKey getSigningKey() {
     byte[] keyBytes = jwtSecret.getBytes(StandardCharsets.UTF_8);
